ADD validations in create place and register screen, ADD fetch and create reports

+++ app/src/main/java/eam/edu/co/applugaresproyectofinal/model/Review.kt
@@ -7,5 +7,5 @@ data class Review(
     val userId: String,
     val subject: String,
     val description: String,
-    val rating: Float
+    val rating: Double
 )
